+++ src/benchstats/common.py
@@ -36,6 +36,24 @@ class ParserBase:
         """
         raise RuntimeError("DERIVED CLASS MUST IMPLEMENT METHOD")
 
+    def getAltDelimiter(self) -> str | None:
+        """Sets comparison mode and returns benchmark <name>|<variant> delimiter.
+
+        If a parser class returns None from this method, this sets two sources
+        comparison mode, i.e. it makes benchstats CLI tool expect that user supplies file1
+        and file2, each containing data_set1 or data_set2 data respectively. The tool
+        must compare data from data_set1 to a matching data in data_set2.
+
+        If a parser class returns a string from this method, this sets a single source
+        comparison method, where the string defines a delimiter substring in a benchmark name
+        that separates the benchmark identifier from the variant identifier. For details,
+        see the docstring of benchstats.compare.compareStats(), especially a description of
+        the alt_delimiter= argument.
+        Note that such types of a parser is only supported for `--file_parser` of `--file1_parser`
+        CLI arguments.
+        """
+        return None
+
 
 def detectExportFormat(export_to, export_fmt):
     assert (export_to is None and export_fmt is None) or (
@@ -45,33 +63,42 @@ def detectExportFormat(export_to, export_fmt):
 
     if export_to is not None and export_fmt is None:
         root, ext = os.path.splitext(export_to)
-        assert ext in [
-            "." + e for e in kAvailableFormats
-        ], f"Unrecognized export file extension '{ext}' of a file in --export_to parameter"
+        assert ext in ["." + e for e in kAvailableFormats], (
+            f"Unrecognized export file extension '{ext}' of a file in --export_to parameter"
+        )
         export_fmt = ext[1:]
 
     return export_fmt
 
 
 class LoggingConsole(rich.console.Console):
-
     # @enum.verify(enum.CONTINUOUS)  # not supported by Py 3.10
     class LogLevel(enum.IntEnum):
-        Debug = (0,)
-        Info = (1,)
-        Warning = (2,)
-        Error = (3,)
-        Failure = (4,)
-        Critical = 5
+        Trace = 0
+        Debug = 1
+        Info = 2
+        Warning = 3
+        Error = 4  # recoverable
+        Failure = 5  # non-recoverable, can continue working
+        Critical = 6  # non-recoverable, must abort
 
     def __init__(self, log_level: LogLevel = LogLevel.Debug, **kwargs):
         assert isinstance(log_level, LoggingConsole.LogLevel)
         self.log_level = log_level
+        self._n_errors: int = 0
         if "emoji" not in kwargs:
             kwargs["emoji"] = False
         if "highlight" not in kwargs:
             kwargs["highlight"] = False
         super().__init__(**kwargs)
+
+    def cleanNumErrors(self) -> int:
+        r = self._n_errors
+        self._n_errors = 0
+        return r
+
+    def getNumErrors(self) -> int:
+        return self._n_errors
 
     def _do_log(self, color: str, lvl: str, *args, **kwargs):
         if "sep" in kwargs:
@@ -81,15 +108,23 @@ class LoggingConsole(rich.console.Console):
             kwargs["sep"] = sep
         return super().print(f"[[{color}]{lvl:4s}[/{color}]]{sep}", *args, **kwargs)
 
+    def will_log(self, level) -> bool:
+        return self.log_level <= level
+
+    def trace(self, *args, **kwargs):
+        if self.log_level > LoggingConsole.LogLevel.Trace:
+            return None
+        return self._do_log("blue", "trce", *args, **kwargs)
+
     def debug(self, *args, **kwargs):
         if self.log_level > LoggingConsole.LogLevel.Debug:
             return None
-        return self._do_log("gray", "dbg", *args, **kwargs)
+        return self._do_log("bright_black", "dbg", *args, **kwargs)
 
     def info(self, *args, **kwargs):
         if self.log_level > LoggingConsole.LogLevel.Info:
             return None
-        return self._do_log("white", "info", *args, **kwargs)
+        return self._do_log("bright_white", "info", *args, **kwargs)
 
     def warning(self, *args, **kwargs):
         if self.log_level > LoggingConsole.LogLevel.Warning:
@@ -97,19 +132,22 @@ class LoggingConsole(rich.console.Console):
         return self._do_log("yellow", "warn", *args, **kwargs)
 
     def error(self, *args, **kwargs):
+        self._n_errors += 1
         if self.log_level > LoggingConsole.LogLevel.Error:
             return None
-        return self._do_log("orange", "Err", *args, **kwargs)
+        return self._do_log("red", "Err", *args, **kwargs)
 
     def failure(self, *args, **kwargs):
+        self._n_errors += 1
         if self.log_level > LoggingConsole.LogLevel.Failure:
             return None
-        return self._do_log("red", "FAIL", *args, **kwargs)
+        return self._do_log("bright_red", "FAIL", *args, **kwargs)
 
     def critical(self, *args, **kwargs):
+        self._n_errors += 1
         if self.log_level > LoggingConsole.LogLevel.Critical:
             return None
-        return self._do_log("magenta", "CRIT", *args, **kwargs)
+        return self._do_log("bright_magenta", "CRIT", *args, **kwargs)
 
 
 def bmNamesTransform(
@@ -123,9 +161,9 @@ def bmNamesTransform(
     assert isinstance(console, LoggingConsole)
     if re_from is None:
         # not nice that it glues func args with CLI args, but ok to simplify err handling
-        assert (
-            re_to is None
-        ), f"--to{set_idx} can only be used when there's a corresponding --from{set_idx}"
+        assert re_to is None, (
+            f"--to{set_idx} can only be used when there's a corresponding --from{set_idx}"
+        )
         return stats
     assert isinstance(re_from, str) and len(re_from) > 0
     if re_to is None: